minor code refactors and style changes
-- added nav bar popping down after scroll amount
