+++ components/faq/faq.tsx
@@ -44,15 +44,15 @@ const faqs = [
 
 const FAQPage = () => {
 	return (
-		<div className="h-auto mb-8 flex flex-col items-start text-white bg-base-100 " id="faq">
-			<div className="w-full desktop:w-5/6">
+		<div className="h-auto mb-8 flex flex-col items-center text-white bg-base-100 " id="faq">
+			<div className="w-full desktop:w-2/3">
 				<h1 className="font-mokoto font-normal text-white text-left text-4xl text-shadow-md pb-4">FAQs</h1>
 			</div>
-			<div className="w-full desktop:w-5/6">
+			<div className="w-full desktop:w-2/3">
 				{faqs.map((faq, index) => (
 					<div
 						key={index}
-						className={`collapse collapse-arrow custom-arrow bg-base-200 p-4 text-2xl border-t-2 ${
+						className={`collapse collapse-arrow custom-arrow bg-base-200 p-1 text-2xl border-t-2 ${
 							index === faqs.length - 1 ? "border-b-2" : ""
 						} border-hackrpi-primary-blue rounded-none`}
 					>
@@ -64,7 +64,7 @@ const FAQPage = () => {
 					</div>
 				))}
 			</div>
-			<div className="w-full desktop:w-5/6 ">
+			<div className="w-full desktop:w-2/3 ">
 				<h2 id="sponsors" className="font-poppins text-2xl text-center pt-10">
 					Feel free to contact us with any other questions at{" "}
 					<a href="mailto:hackrpi@rpi.edu" className="text-hackrpi-primary-blue">
