+++ components/title-components/registration-button.tsx
